Module rusqbin::storage

Defines a Bins trait for storing Requests as well as a default in-memory implementation of that trait for easy testing.

Structs

InMemoryBins
A simple in-memory implementation of Bins.

Enums

DeleteBinStatus
ADT For deleting a bin by id

InsertRequestStatus
ADT for denoting status when inserting a request with a bin id.

Traits

Bins
Trait for storage operations for Requests.

Type Definitions

Bin
A Bin holds a bunch of requests. For now it's just an alias for a Vector.