Trait rusqbin::storage::Bins

pub trait Bins {
    fn create_bin(&mut self) -> BinSummary;

    fn delete_bin(&mut self, id: &Id) -> DeleteBinStatus;

    fn get_bin<'a>(&'a self, id: &'a Id) -> Option<&Bin>;

    fn get_bin_summary(&self, id: &Id) -> Option<BinSummary>;

    fn get_bin_summaries(&self) -> HashMap<Id, BinSummary>;

    fn insert_request(
        &mut self, 
        id: &Id, 
        request: Request
    ) -> InsertRequestStatus;
}

Trait for storage operations for Requests.

Required Methods

fn create_bin(&mut self) -> BinSummary

Returns a BinSummary of a newly-reated bin. The Id in the summary must be unique at the time of creation.

fn delete_bin(&mut self, id: &Id) -> DeleteBinStatus

Delete a bin by Id

fn get_bin<'a>(&'a self, id: &'a Id) -> Option<&Bin>

Get a bin (not just a summary) by Id

fn get_bin_summary(&self, id: &Id) -> Option<BinSummary>

Get a bin summary by Id

fn get_bin_summaries(&self) -> HashMap<Id, BinSummary>

Get bin summaries for all currently-stored bins

fn insert_request(&mut self, id: &Id, request: Request) -> InsertRequestStatus

Insert a request into a Bin using a bin Id.